// packages/vi-lib/src/vctp-parsing/vctp-types.ts
export enum VCTPMainType {
  Unknown = -1,
  Number = 0x0,
  Unit = 0x1,
  Bool = 0x2,
  Blob = 0x3,
  Array = 0x4,
  Cluster = 0x5,
  Block = 0x6,
  Ref = 0x7,
  NumberPointer = 0x8,
  Terminal = 0xF,
  Void = 0x100,
  Value = -2,
}


export enum VCTPType {
  Unknown = -1,
  Void = 0,
  NumberI8  = 0x1,
  NumberI16 = 0x2,
  NumberI32 = 0x3,
  NumberI64 = 0x4,
  NumberU8  = 0x5,
  NumberU16 = 0x6,
  NumberU32 = 0x7,
  NumberU64 = 0x8,
  NumberSGL = 0x9,
  NumberDBL = 0xA,
  NumberXTP = 0xB,
  NumberCSG = 0xC,
  NumberCDB = 0xD,
  NumberCXT = 0xE,

  UnitI8  = 0x11,
  UnitI16 = 0x12,
  UnitI32 = 0x13,
  UnitI64 = 0x14,
  UnitU8  = 0x15,
  UnitU16 = 0x16,
  UnitU32 = 0x17,
  UnitU64 = 0x18,
  UnitSGL = 0x19,
  UnitDBL = 0x1A,
  UnitXTP = 0x1B,
  UnitCSG = 0x1C,
  UnitCDB = 0x1D,
  UnitCXT = 0x1E,

  BoolU16 =	0x20,
  Bool    = 0x21,

  String  = 0x30,
  String2 =	0x31,
  Path    = 0x32,
  Picture = 0x33,
  CString = 0x34,
  PascalString = 0x35,
  DAQChannel   = 0x37,
  SubString    = 0x3F,

  Array        = 0x40,
  ArrayDataPtr = 0x41,
  SubArray     = 0x4F,

  Cluster        = 0x50,
  ClusterVariant = 0x53,
  ClusterData    = 0x54,
  ComplexFixedPt = 0x5E,
  ClusterNumFixPoint = 0x5F,

  Block         = 0x60,
  TypeBlock     = 0x61,
  VoidBlock     = 0x62,
  AlignedBlock  = 0x63,
  RepeatedBlock = 0x64,
  AlignmentMarker = 0x65,

  Refnum = 0x70,

  PointerNumber    = 0x80,
  PointerNumberI8  = 0x81,
  PointerNumberI16 = 0x82,
  PointerNumberI32 = 0x83,
  PointerNumberI64 = 0x84,
  PointerNumberU8  = 0x85,
  PointerNumberU16 = 0x86,
  PointerNumberU32 = 0x87,
  PointerNumberU64 = 0x88,
  PointerNumberSGL = 0x89,
  PointerNumberDBL = 0x8A,
  PointerNumberXTP = 0x8B,
  PointerNumberCSG = 0x8C,
  PointerNumberCDB = 0x8D,
  PointerNumberCXT = 0x8E,

  ArrayInterfc =	0xA0,
  InterfcToData =	0xA1,

  Terminal =  0xF0,
  TypeDef =   0xF1,
  PolyVI =    0xF2,

  EnumValue = -2,
};